Optimized the newm.php page
The breadcrumbs are inserted in the db using one query.
mass_make_thread_breadcrumbs() function added to functions.php. The
recipients id's are now selected based on one foreach checking if the
user exists instead of two. Onld way: one loop for the not existing
users and one for the id of the recipients.